+++ libs/core/src/components/index.ts
@@ -44,3 +44,4 @@ export * from './devfile-header/devfile-header';
 export * from './devfile-starter-projects/devfile-starter-projects';
 export * from './devfile-datalist/devfile-datalist';
 export * from './devfile-codeblock/devfile-codeblock';
+export * from './custom-404/custom-404';
